requirements.txt
pymongo[srv]
requests
pandas
fitz
pillow 
pytesseract
pymupdf
transformers 
accelerate 
torch
python-dotenv